Added placeholder images and prices for each type of clothing. Unable to make shopping cart work and save information across all pages.
