+++ src/app/header/page.tsx
@@ -8,8 +8,8 @@ export default function Header() {
         <div className="flex items-start">  
           <Image src="/globe.svg" alt="website icon" width={50} height={50}/>
           <Link href="/" className="font-mono header">Placeholder</Link>
-          <Link href="/clothing/mens" className="font-mono header">Mens</Link>
-          <Link href="/clothing/womens" className="font-mono header">Womens</Link>
+          <Link href="/clothing/mens" className="font-mono header">Men</Link>
+          <Link href="/clothing/womens" className="font-mono header">Women</Link>
           <Link href="/clothing/kids" className="font-mono header">Kids</Link>
         </div>
         {/* Button Container */}
