+++ src/app/page.tsx
@@ -5,82 +5,24 @@ import Image from "next/image";
 // New Code
 import Link from 'next/link';
 import { useMemo, useState } from "react";
-
-// ---------- Minimal Cart Types ----------
-type Category = "shirts" | "shoes" | "pants" | "other";
-type Product = { id: string; name: string; price: number; category: Category; };
-type CartLine = { productId: string; quantity: number; };
-
-// ---------- Tiny Demo Catalog ----------
-const CATALOG: Product[] = [
-  { id: "s1", name: "Classic Tee",      price: 2200,  category: "shirts" },
-  { id: "p1", name: "Slim Chinos",      price: 4600,  category: "pants"  },
-  { id: "sh1",name: "Everyday Sneakers",price: 6900,  category: "shoes"  },
-  { id: "o1", name: "Crew Socks (3pk)", price: 1400,  category: "other"  },
-];
-
-const formatCurrency = (cents: number) =>
-  (cents / 100).toLocaleString(undefined, { style: "currency", currency: "USD" });
+import { ShoppingCart, CATALOG, formatCurrency } from './cart/page';
 
 export default function Home() {
-  // ---------- Cart State ----------
-  const [cartOpen, setCartOpen] = useState(false);
-  const [cart, setCart] = useState<CartLine[]>([]);
-
-  const productById = useMemo(() => {
-    const m = new Map<string, Product>();
-    for (const p of CATALOG) m.set(p.id, p);
-    return m;
-  }, []);
-
-  const cartWithProducts = useMemo(() => {
-    return cart
-      .map(l => ({ line: l, product: productById.get(l.productId)! }))
-      .filter(x => Boolean(x.product));
-  }, [cart, productById]);
-
-  const subtotal = cartWithProducts.reduce((sum, x) => sum + x.product.price * x.line.quantity, 0);
-  const TAX_RATE = 0.085;
-  const tax = Math.round(subtotal * TAX_RATE);
-  const total = subtotal + tax;
-
-  // ---------- Cart Ops ----------
-  const addToCart = (productId: string, qty = 1) => {
-    setCart(prev => {
-      const i = prev.findIndex(l => l.productId === productId);
-      if (i === -1) return [...prev, { productId, quantity: qty }];
-      const next = [...prev];
-      next[i] = { ...next[i], quantity: next[i].quantity + qty };
-      return next;
-    });
-    setCartOpen(true);
-  };
-
-  const setQuantity = (productId: string, qty: number) => {
-    setCart(prev => prev.map(l => l.productId === productId ? { ...l, quantity: Math.max(1, qty) } : l));
-  };
-
-  const removeFromCart = (productId: string) => {
-    setCart(prev => prev.filter(l => l.productId !== productId));
-  };
-
-  const clearCart = () => setCart([]);
-
-  const checkout = () => {
-    // Placeholder checkout
-    alert(`Order placed!\n\n${JSON.stringify({
-      lines: cartWithProducts.map(({ line, product }) => ({
-        productId: line.productId,
-        name: product.name,
-        unitPrice: product.price,
-        quantity: line.quantity,
-        lineTotal: product.price * line.quantity,
-      })),
-      subtotal, tax, total, placedAt: new Date().toISOString()
-    }, null, 2)}`);
-    clearCart();
-    setCartOpen(false);
-  };
+  const {
+    cartOpen,
+    setCartOpen,
+    cart,
+    cartWithProducts,
+    subtotal,
+    TAX_RATE,
+    tax,
+    total, 
+    addToCart,
+    setQuantity,
+    removeFromCart,
+    clearCart,
+    checkout
+  } = ShoppingCart();
 
   return (
     // Apply the animated background fade here:
@@ -118,7 +60,7 @@ export default function Home() {
         <div id="text section">
           <div id ="title" className="sectitle">
             <Link href="/clothing/mens">
-              <h2>Mens</h2>
+              <h2>Men</h2>
             </Link>
           </div>
           <div id="description" className=" w-2/3 mx-auto">
@@ -137,7 +79,7 @@ export default function Home() {
         <div id="text section">
           <div id ="title" className="sectitle">
             <Link href="/clothing/womens">
-              <h2>Womens</h2>
+              <h2>Women</h2>
             </Link>
           </div>
           <div id="description" className=" w-2/3 mx-auto">
